backend/requirements.txt
fastapi==0.85.1
pydantic==1.10.2
uvicorn==0.18.3
python-dotenv==1.0.0
requests==2.31.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0